Implement ViewPager to replace TabActivity for Mountain Pass Conditions.
ViewPager allows user to swipe left and right to access tab content. Currently
a runtime error is thrown when using MapActivity with the ViewPager so this
may not work.

TabActivity is deprecated, however, a replacement  can be be implemented
by directly using TabHost. May need to look at this if I can't get the
MapActivity to work correctly.